Moved temeprature conversion routines and used them in the mapper
