+++ src/omotes_simulator_core/entities/assets/utils.py
@@ -124,21 +124,3 @@ def sign_output(port_number: int) -> int:
     :return: The multiplication factor for the output.
     """
     return -1 + 2 * (port_number % 2)
-
-
-def celcius_to_kelvin(temperature: float) -> float:
-    """Convert Celcius to Kelvin.
-
-    :param temperature: temperature in C
-    :return: temperature in K
-    """
-    return temperature + 273.15
-
-
-def kelvin_to_celcius(temperature: float) -> float:
-    """Convert Kelvin to Celcius.
-
-    :param temperature: temperature in K
-    :return: temperature in C
-    """
-    return temperature - 273.15
